+++ packages/vocab-tools/src/type.ts
@@ -6,8 +6,10 @@ const HEURISTICS_CONTEXTS: string[] = [
   "https://www.w3.org/ns/activitystreams",
   "https://w3id.org/security/v1",
   "https://w3id.org/security/data-integrity/v1",
+  "https://w3id.org/security/data-integrity/v2",
   "https://www.w3.org/ns/did/v1",
   "https://w3id.org/security/multikey/v1",
+  "https://w3id.org/fep/ef61",
 ];
 
 interface ScalarType {